requirements.txt
hydra-core
omegaconf
torchmetrics==0.6.0
pytorch-lightning>=1.6,<1.7
scikit-learn
tensorboard
torch>=1.11.0
torchvision>=0.12.0
tqdm
h5py
natsort
pandas
timm>=0.5.4
albumentations
